JAMES-1978 Make Subscription tests more generic
With many thanks to:

    Nguyen Thi Mai and Le Thi Huong Lai for their work on Quotas and In memory subscriptions
    Tran Thi & My Linh for their work on TooMuchLines
    txc1996 & Van Thanh For improving readability of SMTPAuthIsSuccessfulTest
    thienan090196 & Tran Thi & My Linh for their tests on RecipientIsLocalTest

From Passerelle Numeriques VietNam
